refactor(client): move interactive-examples hook to telemetry
